Interpolation feature (#2)
* Removed test text file

* Added stopwatch class for perf testing, and also fixed the lambda function in start

* Pushed stuff to my laptop

* Moved to a variable game timer, testing it now

* Tested the variable game clock, it works really well. The old method simply used a Thread.Sleep to run something every ~16.6 milliseconds in order to achieve 60 fps, now we're using a variable game clock and letting the updates happen limitlessly, but always clamping them in a sense by giving a interpolation value. That value basically determines how fast or slow the clock ran between the last execution and this one, and multiplies any value that changes over time by that interpolation value in order to give buttery smooth game updates

+++ src/Utilities/ColorFinder.java
@@ -19,9 +19,9 @@ public class ColorFinder {
 
     }
 
-    public Color getColorInRange(int value, int max) {
+    public Color getColorInRange(double value, double max) {
 
-        double percentage = (double) value / (double) max;
+        double percentage = value / max;
 
         int r = beginColor.getRed() + (int) (percentage * rDifference);
         int g = beginColor.getGreen() + (int) (percentage * gDifference);
